+++ corehq/apps/accounting/mixins.py
@@ -5,8 +5,8 @@ from corehq.apps.accounting.models import CreditLine, Subscription
 from corehq.apps.accounting.tasks import (
     is_subscription_eligible_for_downgrade_process,
 )
-from corehq.apps.accounting.utils import (
-    months_from_date,
+from corehq.apps.accounting.utils import months_from_date
+from corehq.apps.accounting.utils.invoicing import (
     get_unpaid_invoices_over_threshold_by_domain,
 )
 from corehq.apps.users.decorators import get_permission_name
